+++ src/hooks/useAppContext.ts
@@ -1,0 +1,37 @@
+import { createContext, useContext } from "react";
+
+// 커스텀 훅: useAppContext
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface Notification {
+  id: number;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+}
+
+export interface AppContextType {
+  theme: string;
+  toggleTheme: () => void;
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const AppContext = createContext<AppContextType | undefined>(undefined);
+
+// 메인 App 컴포넌트
+
+export const useAppContext = () => {
+  const context = useContext(AppContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
